Don't remove .gitignore when Codeception clen logs

+++ src/Codeception/Util/FileSystem.php
@@ -22,6 +22,9 @@ class FileSystem
                 }
                 rmdir($dir);
             } else {
+                if( basename( (string) $path ) === '.gitignore' ){
+                    continue;
+                }
                 unlink($path->__toString());
             }
         }
